+++ help/docs/releasenotes.md
@@ -16,6 +16,7 @@ This version includes more updates to the scripting engine, including the abilit
 * Fixed the `ignore` and `unignore` commands
 * Fixed the `varexists` expression
 * Added the `unsetvar` command
+* Updated the `interrupt` command to let you pass a specific layer
 
 ## 1.8.61.0 (12/07/2022)